rename package news_list_ex
